Decoupled the DB layer from the feature-layer Issue type
