+++ db/archived-issues.mapper.ts
@@ -1,5 +1,5 @@
 import type { ContributionPlatform } from "@/lib/auth/auth.types";
-import type { Issue } from "@/features/dashboard/dashboard.types";
+import type { Issue } from "@/lib/domain/issues.types";
 
 export type FirestoreArchivedIssue = Issue & {
   platform: ContributionPlatform;
