+++ features/dashboard/shared/issues/hooks/use-archive-issue.hook.ts
@@ -2,8 +2,8 @@
 
 import { useSession } from "next-auth/react";
 import { useProjectIssuesStore } from "../store/project-issues.store";
-import { Issue } from "../../../dashboard.types";
 import { CategorizedProjectIssues } from "../../../dashboard.types";
+import type { Issue } from "@/lib/domain/issues.types";
 
 import { CONSTANTS } from "@/lib/constants";
 import { archiveIssueForPlatform } from "../services/archived-issues-api.service";
