+++ features/dashboard/shared/issues/services/archived-issues-api.service.ts
@@ -1,5 +1,5 @@
 import type { ContributionPlatform } from "@/lib/auth/auth.types";
-import type { Issue } from "@/features/dashboard/dashboard.types";
+import type { Issue } from "@/lib/domain/issues.types";
 
 export async function getArchivedIssuesForPlatform(
   platform: ContributionPlatform,
