+++ features/dashboard/shared/unanswered-issues.tab.tsx
@@ -7,7 +7,8 @@ import { RawIssue } from "@/lib/github/github.types";
 import { LoadingIndicator } from "@/components/layout/loading-indicator";
 import { useLoading } from "@/components/providers/loader-context";
 import { TeamTabs } from "./team-tabs";
-import { CategorizedProjectIssues, Issue } from "../dashboard.types";
+import type { Issue } from "@/lib/domain/issues.types";
+import { CategorizedProjectIssues } from "../dashboard.types";
 import { useProjectIssuesStore } from "./issues/store/project-issues.store";
 import { categorizeIssues } from "./issues/services/categorize-issues.service";
 import type { ContributionPlatform } from "@/lib/auth/auth.types";
