+++ lib/domain/issues.types.ts
@@ -1,0 +1,8 @@
+export interface Issue {
+  issueNumber: number;
+  issueUrl: string;
+  issueTitle: string;
+  isArchived: boolean;
+  lastCommentCreatedAt: string;
+  linkedProject: string;
+}
